MB-1496 issue fixed.
- Public jira url https://wso2.org/jira/browse/MB-1496

We are currently supported wildcard like topicOrSubtopicName.* or topicOrSubtopicName.# and permission model working for subscription like that. If you create subscription like *.topicOrSubtopicName, then it it will cause to permission denied and even you subscribe with admin user, messages are not getting routed to destination. We will improve subscription like *.topicOrSubtopicName in future release.